requirements.txt
pytest==7.4.0
playwright==1.36.0
pytest-playwright==0.3.0
allure-pytest==2.13.2